feat(golden-triangle): dock the priority triangle at the composer (in view, no clipping)
Founder feedback: keep the triangle in view at the composer (don't hide it behind the header Controls menu), and the menu popover clipped off the window edge.

Moves the per-coworker priority OUT of the header posture-menu popover and INTO an in-flow dock just above the message input (CoworkerPriorityDock). In-flow = nothing overflows the window (fixes the clip); persistent = the gradient triangle stays in view where you type. Collapsible (default open) to reclaim space; the collapsed header still shows the colour-graded glyph + preset. Loads + saves THIS coworker's own posture (debounced) — unchanged per-agent behaviour. The superseded header chip (CoworkerPriorityControl) + its test are removed; the dock replaces it. 259 agent+GT tests green; tsc clean.

Process-Spine-Decision: UI placement fix for the per-coworker posture; additive, no contract change. UX-Fit-Decision: one persistent in-flow control at the composer replacing a clipping popover; default-visible triangle, collapsible; decided on merits (human_cognitive_load axis has no kernel scorer).
